+++ include/Flux/Graphics/TextSystem.hpp
@@ -96,7 +96,7 @@ class TextSystem {
 
     virtual std::uint32_t resolveFontId(std::string_view fontFamily, float weight, bool italic) = 0;
 
-    virtual std::vector<std::uint8_t> rasterizeGlyph(std::uint32_t fontId, std::uint16_t glyphId, float size,
+    virtual std::vector<std::uint8_t> rasterizeGlyph(std::uint32_t fontId, std::uint32_t glyphId, float size,
                                                      std::uint32_t &outWidth, std::uint32_t &outHeight,
                                                      Point &outBearing) = 0;
